+++ buildVars.py
@@ -16,28 +16,29 @@ def _(arg):
 # Add-on information variables
 addon_info = {
 	# add-on Name/identifier, internal for NVDA
-	"addon_name": "emulateNumpadMode",
+	"addon_name": "navigateTheNumpadWay",
 	# Add-on summary, usually the user visible name of the addon.
 	# Translators: Summary for this add-on
 	# to be shown on installation and add-on information found in Add-ons Manager.
-	"addon_summary": _("Emulate Numpad Mode"),
+	"addon_summary": _("Navigate The Numpad Way"),
 	# Add-on description
 	# Translators: Long description to be shown for this add-on on add-on information from add-ons manager
-	"addon_description": _("""Emulate numpad key navigation feature
+	"addon_description": _("""Emulate numpad keys commands or navigation feature
 on a portion of the main keyboard.
+You can choose this portion or field, from the setting dialog of the addon.
 Default shortcut to toggle between normal and numpad mode in the addon is Shift+NVDA+F4, and you can modify it from input gestures dialog in NVDA."""),
 	# version
-	"addon_version": "0.9",
+	"addon_version": "1.0",
 	# Author(s)
 	"addon_author": "ibrahim hamadeh <ibra.hamadeh@hotmail.com>",
 	# URL for the add-on documentation support
-	"addon_url": "https://github.com/ibrahim-s/emulateNumpadMode",
+	"addon_url": "https://github.com/ibrahim-s/navigateTheNumpadWay",
 	# Documentation file name
 	"addon_docFileName": "readme.html",
 	# Minimum NVDA version supported (e.g. "2018.3.0", minor version is optional)
 	"addon_minimumNVDAVersion": "2019.3.0",
 	# Last NVDA version supported/tested (e.g. "2018.4.0", ideally more recent than minimum version)
-	"addon_lastTestedNVDAVersion": "2023.1.0",
+	"addon_lastTestedNVDAVersion": "2024.1.0",
 	# Add-on update channel (default is None, denoting stable releases,
 	# and for development releases, use "dev".)
 	# Do not change unless you know what you are doing!
@@ -52,7 +53,7 @@ Default shortcut to toggle between normal and numpad mode in the addon is Shift+
 # pythonSources = ["addon/globalPlugins/*.py"]
 # For more information on SCons Glob expressions please take a look at:
 # https://scons.org/doc/production/HTML/scons-user/apd.html
-pythonSources = ["addon/globalPlugins/emulateNumpadMode/*.py"]
+pythonSources = ["addon/globalPlugins/navigateTheNumpadWay/*.py"]
 
 # Files that contain strings for translation. Usually your python sources
 i18nSources = pythonSources + ["buildVars.py"]