+++ changelog.md
@@ -1,4 +1,4 @@
-## Changes for 0.9 .
+### Changes for 1.0 ###
 
-*	Fix a bug in numbers mode, related to position of number 0, in third and fourth forms.
-*	Update last tested version, thus making the addon compatible with NVDA 2023.1.
+*	Rename the addon to navigate the numpad way, for there are addons, that their names very close to the previous names of the addon, thus renaming the addon make things clearer. 
+*	Update last tested version, thus making the addon compatible with NVDA 2024.1.